undo rename - persistence has an an e in the last syllable
