+++ mobile-save-for-later-user-deletion/src/main/scala/com/gu/sfl/userdeletion/UserDeletionLambda.scala
@@ -3,7 +3,7 @@ package com.gu.sfl.userdeletion
 import com.amazonaws.services.lambda.runtime.events.SQSEvent
 import com.gu.sfl.Logging
 import com.gu.sfl.lib.Jackson.mapper
-import com.gu.sfl.persistance.PersistanceConfig
+import com.gu.sfl.persistence.PersistenceConfig
 import com.gu.sfl.userdeletion.model.UserDeleteMessage
 import coml.gu.sfl.userdeletion.db.SflDynamoDb
 
@@ -16,7 +16,7 @@ object UserDeletionLambda extends Logging {
   val saveForLaterApp = Option(System.getenv("SaveForLaterApp")).getOrElse(sys.error("No main app name configured"))
   val stage = Option(System.getenv("Stage")).getOrElse(sys.error("No main stage configured"))
 
-  val sflDyanamo = new SflDynamoDb(PersistanceConfig(saveForLaterApp, stage))
+  val sflDyanamo = new SflDynamoDb(PersistenceConfig(saveForLaterApp, stage))
 
   def handler(sQSEvent: SQSEvent) {
     val messages = sQSEvent.getRecords.asScala.map(mes => mes).toList
