+++ mobile-save-for-later/src/main/scala/com/gu/sfl/lambda/FetchArticlesLambda.scala
@@ -5,7 +5,7 @@ import com.gu.sfl.Logging
 import com.gu.sfl.controller.FetchArticlesController
 import com.gu.sfl.identity.{IdentityConfig, IdentityServiceImpl}
 import com.gu.sfl.lib.GlobalHttpClient
-import com.gu.sfl.persistance.{PersistanceConfig, SavedArticlesPersistanceImpl}
+import com.gu.sfl.persistence.{PersistenceConfig, SavedArticlesPersistenceImpl}
 import com.gu.sfl.savedarticles.FetchSavedArticlesImpl
 import com.gu.sfl.lib.SsmConfig
 
@@ -23,7 +23,7 @@ object FetchArticlesLambda extends Logging {
           new FetchArticlesController(
             new FetchSavedArticlesImpl(
               new IdentityServiceImpl(IdentityConfig(ssmConfig.config.getString("identity.apiHost")),GlobalHttpClient.defaultHttpClient),
-              new SavedArticlesPersistanceImpl( PersistanceConfig(awsIdentity.app, awsIdentity.stage) )
+              new SavedArticlesPersistenceImpl( PersistenceConfig(awsIdentity.app, awsIdentity.stage) )
             )
           )
         case _ => throw new IllegalStateException("Unable to retrieve configuration")
