+++ mobile-save-for-later/src/main/scala/com/gu/sfl/lambda/SaveArticlesLambda.scala
@@ -5,7 +5,7 @@ import com.gu.sfl.Logging
 import com.gu.sfl.controller.SaveArticlesController
 import com.gu.sfl.identity.{IdentityConfig, IdentityServiceImpl}
 import com.gu.sfl.lib.{GlobalHttpClient, SavedArticlesMergerConfig, SavedArticlesMergerImpl}
-import com.gu.sfl.persistance.{PersistanceConfig, SavedArticlesPersistanceImpl}
+import com.gu.sfl.persistence.{PersistenceConfig, SavedArticlesPersistenceImpl}
 import com.gu.sfl.savedarticles.UpdateSavedArticlesImpl
 import com.gu.sfl.lib.SsmConfig
 
@@ -24,7 +24,7 @@ object SaveArticlesLambda extends Logging {
             new UpdateSavedArticlesImpl(
               new IdentityServiceImpl(IdentityConfig(ssmConfig.config.getString("identity.apiHost")), GlobalHttpClient.defaultHttpClient),
               new SavedArticlesMergerImpl( SavedArticlesMergerConfig(ssmConfig.config.getInt("savedarticle.limit")),
-                new SavedArticlesPersistanceImpl( PersistanceConfig(awsIdentity.app, awsIdentity.stage) )
+                new SavedArticlesPersistenceImpl( PersistenceConfig(awsIdentity.app, awsIdentity.stage) )
               )
             )
           )
