+++ mobile-save-for-later/src/test/scala/com/gu/sfl/lib/ArticleMergeSpecification.scala
@@ -4,7 +4,7 @@ import java.time.LocalDateTime
 
 import com.gu.sfl.exception.SavedArticleMergeError
 import com.gu.sfl.model.{SavedArticle, SavedArticles}
-import com.gu.sfl.persistance.SavedArticlesPersistanceImpl
+import com.gu.sfl.persistence.SavedArticlesPersistenceImpl
 import org.specs2.mock.Mockito
 import org.specs2.mutable.Specification
 import org.specs2.specification.Scope
@@ -147,7 +147,7 @@ class ArticleMergeSpecification extends Specification with Mockito  {
   }
 
   trait Setup extends Scope {
-    val savedArticlesPersistence = mock[SavedArticlesPersistanceImpl]
+    val savedArticlesPersistence = mock[SavedArticlesPersistenceImpl]
     val savedArticlesMerger = new SavedArticlesMergerImpl(SavedArticlesMergerConfig(20), savedArticlesPersistence)
   }
